# docs/openapi.json
{
  "components": {
    "schemas": {
      "Actor": {
        "description": "Who made a change: whoever holds an API key, or Charon itself.",
        "properties": {
          "key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null for Charon itself and the bootstrap key.",
            "title": "Key Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Actor",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Field-level problems, for request validation errors.",
            "title": "Details"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What to do about it, in plain language, when Charon knows.",
            "title": "Hint"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "description": "Whether sending the same request again later may succeed.",
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "MatchType": {
        "enum": [
          "glob",
          "regex"
        ],
        "title": "MatchType",
        "type": "string"
      },
      "PreviewRequest": {
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleSpec"
              },
              {
                "type": "null"
              }
            ],
            "description": "Preview this unsaved rule instead."
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Force this saved rule.",
            "title": "Rule Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "PreviewRequest",
        "type": "object"
      },
      "PreviewView": {
        "properties": {
          "final_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when no rule applies.",
            "title": "Final Path"
          },
          "new_name": {
            "title": "New Name",
            "type": "string"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The saved rule that applied, if any.",
            "title": "Rule Id"
          }
        },
        "required": [
          "rule_id",
          "new_name",
          "final_path"
        ],
        "title": "PreviewView",
        "type": "object"
      },
...
      "RenameOp": {
        "enum": [
          "replace",
          "regex_replace"
        ],
        "title": "RenameOp",
        "type": "string"
      },
      "RenameStep": {
        "properties": {
          "find": {
            "minLength": 1,
            "title": "Find",
            "type": "string"
          },
          "op": {
            "$ref": "#/components/schemas/RenameOp"
          },
          "replace": {
            "default": "",
            "title": "Replace",
            "type": "string"
          }
        },
        "required": [
          "op",
          "find"
        ],
        "title": "RenameStep",
        "type": "object"
      },
      "ReorderRulesRequest": {
        "properties": {
          "ids": {
            "description": "Every rule's id, in the new order, highest priority first.",
            "items": {
              "type": "string"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "ReorderRulesRequest",
        "type": "object"
      },
...
      "Rule": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Actor"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "default": "",
            "description": "Why the rule exists, in plain words.",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "match_type": {
            "$ref": "#/components/schemas/MatchType",
            "default": "glob"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pattern": {
            "minLength": 1,
            "title": "Pattern",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RenameStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "updated_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Actor"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "default": 1,
            "description": "Goes up by one on every change.",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "pattern",
          "destination",
          "id",
          "created_at"
        ],
        "title": "Rule",
        "type": "object"
      },
      "RuleSpec": {
        "description": "A rule as supplied by a client, before it has an identity.",
        "properties": {
          "description": {
            "default": "",
            "description": "Why the rule exists, in plain words.",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "match_type": {
            "$ref": "#/components/schemas/MatchType",
            "default": "glob"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pattern": {
            "minLength": 1,
            "title": "Pattern",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RenameStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "name",
          "pattern",
          "destination"
        ],
        "title": "RuleSpec",
        "type": "object"
      },
      "RuleUpdate": {
        "properties": {
          "description": {
            "default": "",
            "description": "Why the rule exists, in plain words.",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "match_type": {
            "$ref": "#/components/schemas/MatchType",
            "default": "glob"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pattern": {
            "minLength": 1,
            "title": "Pattern",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "title": "Priority",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RenameStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The version you loaded. If the rule has changed since, the save fails with 409 rule_changed instead of overwriting that change. Omit to overwrite.",
            "title": "Version"
          }
        },
        "required": [
          "name",
          "pattern",
          "destination"
        ],
        "title": "RuleUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Charon",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rules": {
      "get": {
        "description": "Every rule, in the order they are tried.",
        "operationId": "listRules",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Rule"
                  },
                  "title": "Response Listrules",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Rules",
        "tags": [
          "rules"
        ]
      },
      "post": {
        "operationId": "createRule",
        "parameters": [
          {
            "description": "Any unique string, e.g. a UUID. Repeating a request with the same key returns what the first one created (with status 200) instead of creating it again. Keys belong to the API key that sent them and are remembered for a day.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any unique string, e.g. a UUID. Repeating a request with the same key returns what the first one created (with status 200) instead of creating it again. Keys belong to the API key that sent them and are remembered for a day.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            },
            "description": "Repeated request"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflicts with the resource's current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Rule",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/rules/preview": {
      "post": {
        "description": "Dry run: what `name` would become, under the saved rules or an unsaved draft `rule`.",
        "operationId": "previewRule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Preview Rule",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/rules/reorder": {
      "post": {
        "description": "Put every rule in a new order at once; priorities are re-spaced 10, 20, 30\u2026\n\nFails with 409 rules_changed if `ids` isn't exactly the current set of rules.",
        "operationId": "reorderRules",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderRulesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Rule"
                  },
                  "title": "Response Reorderrules",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflicts with the resource's current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Reorder Rules",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/rules/{rule_id}": {
      "delete": {
        "operationId": "deleteRule",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Rule",
        "tags": [
          "rules"
        ]
      },
      "get": {
        "operationId": "getRule",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Rule",
        "tags": [
          "rules"
        ]
      },
      "put": {
        "description": "Replace a rule. Send the `version` you loaded (or If-Match) to refuse a newer edit.",
        "operationId": "updateRule",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "description": "The rule's ETag, e.g. \"3\": an alternative to sending `version`.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The rule's ETag, e.g. \"3\": an alternative to sending `version`.",
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, invalid or revoked API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflicts with the resource's current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update Rule",
        "tags": [
          "rules"
        ]
      }
    }
  }
}